refactor: move things around in common module
