+++ common/src/main/java/com/badbones69/crazycrates/api/enums/Properties.java
@@ -3,12 +3,12 @@ package com.badbones69.crazycrates.api.enums;
 import ch.jalu.configme.configurationdata.ConfigurationData;
 import ch.jalu.configme.properties.Property;
 import ch.jalu.configme.resource.PropertyReader;
-import com.badbones69.crazycrates.platform.config.impl.ConfigKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.CommandKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.CrateKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.ErrorKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.MiscKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.PlayerKeys;
+import com.badbones69.crazycrates.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.config.impl.messages.CommandKeys;
+import com.badbones69.crazycrates.config.impl.messages.CrateKeys;
+import com.badbones69.crazycrates.config.impl.messages.ErrorKeys;
+import com.badbones69.crazycrates.config.impl.messages.MiscKeys;
+import com.badbones69.crazycrates.config.impl.messages.PlayerKeys;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
