+++ common/src/main/java/com/badbones69/crazycrates/config/ConfigManager.java
@@ -1,17 +1,17 @@
-package com.badbones69.crazycrates.platform.config;
+package com.badbones69.crazycrates.config;
 
 import ch.jalu.configme.SettingsManager;
 import ch.jalu.configme.SettingsManagerBuilder;
 import ch.jalu.configme.resource.YamlFileResourceOptions;
-import org.jetbrains.annotations.ApiStatus;
-import com.badbones69.crazycrates.platform.config.impl.ConfigKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.CommandKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.CrateKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.ErrorKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.MiscKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.PlayerKeys;
-import com.badbones69.crazycrates.platform.config.migrate.ConfigMigration;
-import com.badbones69.crazycrates.platform.config.migrate.LocaleMigration;
+import com.badbones69.crazycrates.config.migrate.ConfigMigration;
+import com.badbones69.crazycrates.config.migrate.LocaleMigration;
+import com.badbones69.crazycrates.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.config.impl.messages.CommandKeys;
+import com.badbones69.crazycrates.config.impl.messages.CrateKeys;
+import com.badbones69.crazycrates.config.impl.messages.ErrorKeys;
+import com.badbones69.crazycrates.config.impl.messages.MiscKeys;
+import com.badbones69.crazycrates.config.impl.messages.PlayerKeys;
+
 import java.io.File;
 
 public class ConfigManager {
