+++ common/src/main/java/com/badbones69/crazycrates/config/migrate/ConfigMigration.java
@@ -1,4 +1,4 @@
-package com.badbones69.crazycrates.platform.config.migrate;
+package com.badbones69.crazycrates.config.migrate;
 
 import ch.jalu.configme.configurationdata.ConfigurationData;
 import ch.jalu.configme.migration.PlainMigrationService;
