+++ paper/src/main/java/com/badbones69/crazycrates/api/enums/Messages.java
@@ -7,16 +7,16 @@ import com.ryderbelserion.vital.enums.Support;
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.Nullable;
-import com.badbones69.crazycrates.platform.config.ConfigManager;
-import com.badbones69.crazycrates.platform.config.impl.messages.CommandKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.CrateKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.ErrorKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.MiscKeys;
-import com.badbones69.crazycrates.platform.config.impl.messages.PlayerKeys;
+import com.badbones69.crazycrates.config.ConfigManager;
+import com.badbones69.crazycrates.config.impl.messages.CommandKeys;
+import com.badbones69.crazycrates.config.impl.messages.CrateKeys;
+import com.badbones69.crazycrates.config.impl.messages.ErrorKeys;
+import com.badbones69.crazycrates.config.impl.messages.MiscKeys;
+import com.badbones69.crazycrates.config.impl.messages.PlayerKeys;
 import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
-import com.badbones69.crazycrates.platform.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.config.impl.ConfigKeys;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
