+++ paper/src/main/java/com/badbones69/crazycrates/tasks/BukkitUserManager.java
@@ -3,8 +3,8 @@ package com.badbones69.crazycrates.tasks;
 import ch.jalu.configme.SettingsManager;
 import com.badbones69.crazycrates.api.enums.Files;
 import com.badbones69.crazycrates.api.utils.ItemUtils;
-import com.badbones69.crazycrates.platform.config.ConfigManager;
-import com.badbones69.crazycrates.platform.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.config.ConfigManager;
+import com.badbones69.crazycrates.config.impl.ConfigKeys;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
 import com.ryderbelserion.vital.files.yaml.FileManager;
 import com.ryderbelserion.vital.util.builders.ItemBuilder;
