+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/other/quadcrates/QuadCrateManager.java
@@ -9,8 +9,8 @@ import org.bukkit.Material;
 import org.bukkit.Particle;
 import org.bukkit.SoundCategory;
 import org.bukkit.plugin.java.JavaPlugin;
-import com.badbones69.crazycrates.platform.config.ConfigManager;
-import com.badbones69.crazycrates.platform.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.config.ConfigManager;
+import com.badbones69.crazycrates.config.impl.ConfigKeys;
 import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
